+++ src/average.c
@@ -4,8 +4,8 @@
 /// Excercise 6
 /// Returns the average of an array
 /// Pre: n>0, list[0...n-1] is defined
-int average(int list[], int n) {
-  int avg = 0;
+double average(int list[], int n) {
+  double avg = 0;
   assert(n > 0);
   //...
   return avg;
